Create last delted words and suggestions with undo option

+++ Web/database/migrations/2021_01_16_112316_create_words_table.php
@@ -17,7 +17,7 @@ class CreateWordsTable extends Migration
             $table->id();
             $table->string('word')->unique();
             $table->string('translation');
-            $table->bigInteger('user_id');
+            $table->integer('user_id');
             $table->timestamps();
         });
     }
